requirements.txt
# Web框架
fastapi==0.104.1
uvicorn==0.24.0

# 模板和静态文件
jinja2==3.1.2

# 文件上传
python-multipart==0.0.6
aiofiles==23.2.1

# 数据库
mysql-connector-python==8.2.0

# 图像处理
Pillow==10.0.1
opencv-python==4.8.1.78

# OCR引擎 - 在Dockerfile中安装
# PaddlePaddle和PaddleOCR将在构建时安装


# 数据处理
numpy==1.24.4
pandas==1.5.3

# HTTP请求
requests==2.31.0

# 系统监控
psutil==5.9.6

# 环境配置
python-dotenv==1.0.0

# MRZ处理（如果需要）
mrz

# 安全和认证（可选）
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 